+++ core/src/mast/debuginfo/mod.rs
@@ -126,10 +126,8 @@ impl DebugInfo {
 
     /// Creates an empty [DebugInfo] with valid CSR structures for N nodes.
     pub fn empty_for_nodes(num_nodes: usize) -> Self {
-        let mut node_indptr_for_op_idx = IndexVec::new();
-        for _ in 0..=num_nodes {
-            let _ = node_indptr_for_op_idx.push(0);
-        }
+        let node_indptr_for_op_idx = IndexVec::try_from(vec![0; num_nodes + 1])
+            .expect("num_nodes should not exceed u32::MAX");
 
         let op_decorator_storage =
             OpToDecoratorIds::from_components(Vec::new(), Vec::new(), node_indptr_for_op_idx)